requirements.txt
olefile==0.44
Pillow==4.2.1
pycrypto==2.6.1
Flask==0.12.2
bitarray==0.8.1
simple-crypt==4.1.7
